+++ src/dawn_native/d3d12/DeviceD3D12.h
@@ -69,7 +69,8 @@ namespace dawn_native { namespace d3d12 {
 
         void ExecuteCommandLists(std::initializer_list<ID3D12CommandList*> commandLists);
 
-        ResultOrError<std::unique_ptr<StagingBufferBase>> CreateStagingBuffer(size_t size) override;
+        ResultOrError<std::unique_ptr<StagingBufferBase>> CreateStagingBuffer(
+            uint64_t size) override;
         MaybeError CopyFromStagingToBuffer(StagingBufferBase* source,
                                            uint64_t sourceOffset,
                                            BufferBase* destination,
