+++ src/dawn_native/d3d12/StagingBufferD3D12.h
@@ -24,7 +24,7 @@ namespace dawn_native { namespace d3d12 {
 
     class StagingBuffer : public StagingBufferBase {
       public:
-        StagingBuffer(size_t size, Device* device);
+        StagingBuffer(uint64_t size, Device* device);
         ~StagingBuffer();
 
         ID3D12Resource* GetResource() const;
